refactored code so that there is a single store not multiple store folders Keeping all the slices in a single folder itself
