+++ src/widget/Features/actionBar/ActionBar.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import HoverComponent from '../common/components/HoverComponent';
 import { themeColors } from '../common/utils/colors';
-import { setChatInterfaceVisible } from '../store/uiVisibilitySlice';
+import { setChatInterfaceVisible } from '../../store/slices/uiVisibilitySlice';
 
 const ActionBar = () => {
   const dispatch = useDispatch();
